1/requirements.txt
flask
dnspython
gunicorn